chore: modularize data model tests
