+++ bc_obps/registration/tests/constants.py
@@ -1,0 +1,18 @@
+OPERATOR_FIXTURE = ("mock/operator.json",)
+USER_FIXTURE = ("mock/user.json",)
+ADDRESS_FIXTURE = ("mock/address.json",)
+OPERATION_FIXTURE = ("mock/operation.json",)
+CONTACT_FIXTURE = ("mock/contact.json",)
+DOCUMENT_FIXTURE = ("mock/document.json",)
+BC_OBPS_REGULATED_OPERATION_FIXTURE = ("mock/bc_obps_regulated_operation.json",)
+FACILITY_FIXTURE = ("mock/facility.json",)
+
+
+TIMESTAMP_COMMON_FIELDS = [
+    ("created_at", "created at", None, None),
+    ("created_by", "created by", None, None),
+    ("updated_at", "updated at", None, None),
+    ("updated_by", "updated by", None, None),
+    ("archived_at", "archived at", None, None),
+    ("archived_by", "archived by", None, None),
+]
