+++ bc_obps/registration/tests/models/test_contact.py
@@ -1,0 +1,34 @@
+from common.tests.utils.helpers import BaseTestCase
+from registration.models import (
+    Document,
+    Contact,
+)
+from registration.tests.constants import ADDRESS_FIXTURE, CONTACT_FIXTURE, DOCUMENT_FIXTURE, USER_FIXTURE, TIMESTAMP_COMMON_FIELDS
+
+
+class ContactModelTest(BaseTestCase):
+    fixtures = [ADDRESS_FIXTURE, USER_FIXTURE, CONTACT_FIXTURE, DOCUMENT_FIXTURE]
+
+    @classmethod
+    def setUpTestData(cls):
+        cls.test_object = Contact.objects.get(id=1)
+        cls.test_object.documents.set([Document.objects.get(id=1), Document.objects.get(id=2)])
+        cls.field_data = [
+            *TIMESTAMP_COMMON_FIELDS,
+            ("id", "ID", None, None),
+            ("first_name", "first name", 1000, None),
+            ("last_name", "last name", 1000, None),
+            ("position_title", "position title", 1000, None),
+            ("address", "address", None, None),
+            ("email", "email", 254, None),
+            (
+                "phone_number",
+                "phone number",
+                None,
+                None,
+            ),  # Replace None with the actual max length if available
+            ("business_role", "business role", None, None),
+            ("documents", "documents", None, None),
+            ("operators", "operator", None, None),
+            ("operations", "operation", None, None),
+        ]
