+++ bc_obps/registration/tests/models/test_parent_operator.py
@@ -1,0 +1,30 @@
+from common.tests.utils.helpers import BaseTestCase
+from registration.tests.constants import ADDRESS_FIXTURE, CONTACT_FIXTURE, OPERATOR_FIXTURE, TIMESTAMP_COMMON_FIELDS, USER_FIXTURE
+from registration.tests.utils.bakers import parent_operator_baker
+
+
+class ParentOperatorModelTest(BaseTestCase):
+    fixtures = [
+        USER_FIXTURE,
+        ADDRESS_FIXTURE,
+        CONTACT_FIXTURE,
+        OPERATOR_FIXTURE,
+    ]
+
+    @classmethod
+    def setUpTestData(cls):
+        cls.test_object = parent_operator_baker()
+        cls.field_data = [
+            *TIMESTAMP_COMMON_FIELDS,
+            ("id", "ID", None, None),
+            ("child_operator", "child operator", None, None),
+            ("operator_index", "operator index", None, None),
+            ("legal_name", "legal name", 1000, None),
+            ("trade_name", "trade name", 1000, None),
+            ("cra_business_number", "cra business number", None, None),
+            ("bc_corporate_registry_number", "bc corporate registry number", None, None),
+            ("business_structure", "business structure", None, None),
+            ("website", "website", 200, None),
+            ("physical_address", "physical address", None, None),
+            ("mailing_address", "mailing address", None, None),
+        ]
